+++ pysnow/__init__.py
@@ -2,10 +2,10 @@
 
 # Public API
 from pysnow.client import Client
-from pysnow.oauth_client import OAuthClient
-from pysnow.query import QueryBuilder
-from pysnow.request import Request
 from pysnow.legacy_request import LegacyRequest
+from pysnow.oauth_client import OAuthClient
+from pysnow.resource import Resource
+from pysnow.query_builder import QueryBuilder
 
 __author__ = "Robert Wikman <rbw@vault13.org>"
 __version__ = "0.6.0"
