+++ pysnow/report.py
@@ -1,0 +1,2 @@
+# -*- coding: utf-8 -*-
+
